requirements.txt
aiohttp>=3.7.4.post0,<4
pycryptodome>=3.11.0
requests>=2.16.0,<3.0.0
typing-extensions>=3.7.4.1,<4;python_version<'3.8'
